% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{find_peaks}
\alias{find_peaks}
\title{Funkcja do szukania pików fluorescencji w danych z time lapse}
\usage{
find_peaks(
  ramka,
  s = 2,
  m = FALSE,
  procent = 1,
  threshold = 10,
  back = FALSE,
  lapse = 10,
  filter_local = FALSE,
  filter_local_int = 1.1,
  filter_local_width = 2,
  ...
)
}
\arguments{
\item{ramka}{tabela z danymi (kolumna 1 - odległości, 2 - intensywność fluorescencji)}

\item{s}{sigma - dokładność dopasowania pików, im wyższa tym mniej znajdzie}

\item{m}{jak ma wygładzać dane, przy FALSE znajduje więcej pików}

\item{procent}{ile tła ma odjąć, rozsądny zakres to od 1(całość) do 0.01 (1 procent)}

\item{threshold}{powyżej jakiego poziomu ma zaznaczać piki, procent wysokości najwyższego piku}

\item{back}{lepiej zostawić FALSE}

\item{lapse}{czas pomiędzy klatkami}

\item{...}{}
}
\value{
ramkę danych zawierającą pozycje wszystkich pików razem z czasem i intensywnośćią fluorescencji
}
\description{
Funkcja do szukania pików fluorescencji w danych z time lapse
}
