% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{plot_find_peaks}
\alias{plot_find_peaks}
\title{Plots results form find_peaks() 
Visualizes indetified maxima}
\usage{
plot_find_peaks(dane_raw, dane_find)
}
\arguments{
\item{dane_raw}{fluorescence profiles data}

\item{dane_find}{data generated by find_peaks}
}
\value{
ggplot of fluorescence maxima
}
\description{
Plots results form find_peaks() 
Visualizes indetified maxima
}
