% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{plot_kymograph_find_peaks}
\alias{plot_kymograph_find_peaks}
\title{Rysuje kymograf fluorescencji na podstawie profilu, zaznacza znalezione maksima}
\usage{
plot_kymograph_find_peaks(
  dane_raw,
  dane_find,
  odwroc = TRUE,
  pokaz = TRUE,
  color_point,
  color_gradient,
  lapse
)
}
\arguments{
\item{dane_raw}{profile fluorescencji}

\item{dane_find}{wynik find_peaks}

\item{odwroc}{gdzie jest "góra" komórki}

\item{pokaz}{czy pokazać maksima}

\item{color_point}{kolor punktów}

\item{color_gradient}{kolor gradientu, zawsze drugi kolor jest czarny}

\item{lapse}{czas pomiędzy klatkami}
}
\description{
Rysuje kymograf fluorescencji na podstawie profilu, zaznacza znalezione maksima
}
