% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{plot_peaks_ridges}
\alias{plot_peaks_ridges}
\title{Title}
\usage{
plot_peaks_ridges(
  data,
  scale = "osobno",
  gradient = TRUE,
  skala = 2,
  reverse = FALSE
)
}
\arguments{
\item{data}{raw fluorescence plots data}

\item{scale}{should data be normalized for each plot individually}

\item{gradient}{plot color gradient? (scale viridis)}

\item{skala}{passed to geom_ridgeline - how scale plots (1 do not overplot)}

\item{reverse}{should plot be reversed (is 0 the tip or the bottom of the hyphae?)}
}
\value{
ggplot
}
\description{
Uses geom_ridgeline to plot all fluorescence profiles on one plot
}
