% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{plot_scheme_find_peaks}
\alias{plot_scheme_find_peaks}
\title{Rysuje schemat komórki na podstawie wyniku find_peaks}
\usage{
plot_scheme_find_peaks(dane_find, odwroc = TRUE, color_point = "red")
}
\arguments{
\item{dane_find}{wynik funkcji find_peaks}

\item{odwroc}{gdzie jest "góra" komórki}
}
\value{
schematyczny wykres komórki
}
\description{
Rysuje schemat komórki na podstawie wyniku find_peaks
}
