% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waves}
\alias{waves}
\title{Sample wave height data}
\format{A data frame with 50 rows and 2 variables:
\describe{
  \item{date_time}{date and time, in as POSIXct}
  \item{height_meters}{wave height, in meters}
  }
  ...}
\usage{
waves
}
\description{
A dataset containing wave heights as timeseries
}
\keyword{datasets}
