% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPeaks.R
\name{show.peaks}
\alias{show.peaks}
\title{Plot peaks}
\usage{
show.peaks(x, threshold = 0, ...)
}
\arguments{
\item{x}{Path to the input file}

\item{threshold}{the value above which to calculate peaks}
}
\value{
A plot of the peaks at the specified threshold
}
\description{
This function plots the peak values above the entered threshold argument.
}
