% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{wykres_str}
\alias{wykres_str}
\title{Title}
\usage{
wykres_str(wynik, int = c("int_nor", "int_raw"), zaznacz = FALSE, size = TRUE)
}
\arguments{
\item{wynik}{}

\item{int}{}

\item{zaznacz}{}

\item{size}{}
}
\value{

}
\description{
Title
}
